fix(locator): :bug: fix wrong input points used in clustering
The original points to be used in clustering are pixel points with depths in the lidar coordinate, which is not suitable because the distances between points in different coordinates are not in the same scale. The fix transfers the points into the lidar coordinate so that the scale will become the same.